Add the ability to limit repository releases
This adjusts the directory builder API such that the maximum number
of releases to appear for each package in the generated repository
can be specified. The API also returns the APK files that were not
used in the generation of the resulting repository.

Affects: https://github.com/AULFA/repomaker/issues/1