Add an option to allow the manager to delete APKs
This builds on the previous changes that added the ability for the
directory builder to state which APK files it _did not_ use in the
generation of a repository. The manager now has a configuration option
that allows it to delete APK files that were not used. This should
ensure that APK repositories never grow too large.

Fix: https://github.com/AULFA/repomaker/issues/1